requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Flask>=3.0.0
APScheduler>=3.10.0
readability-lxml>=0.8.1
trafilatura>=1.12.0
lxml_html_clean>=0.4.0
markdown>=3.5.0
